Added support for updating files manually

+++ tools/resources.py
@@ -1,5 +1,6 @@
 import glob
 import json
+from os.path import isfile
 
 from .github import get_commits, get_releases
 
@@ -28,42 +29,55 @@ def update_list():
         opened.write("\n")
 
 
+def update_version(file):
+    """
+    Updates the selected versions from the file.
+    """
+    # If the file does not exists, notify the user and return
+    if not isfile(file):
+        print(f"The file {file} does not exists!")
+        return
+
+    # Open it up and load the contents
+    with open(file, "r+") as opened:
+        # Get the information of the resource
+        data = json.load(opened)
+
+        # Create some shortcuts for the objects
+        params = data["update"].get("parameters", {})
+
+        # If the file uses GitHub Releases for the updates
+        if data["update"]["type"] == 1:
+            # Create a new list of versions
+            data["versions"] = []
+
+            # If we need to fetch a specific number of commits
+            if "commits" in params:
+                # Add those versions from the commits
+                data["versions"].extend(get_commits(**params))
+
+            # If we don't have the update from releases disabled
+            if not ("skip_releases" in params and params["skip_releases"]):
+                # Request the list of releases to GitHub and save them
+                data["versions"] = get_releases(**params)
+
+        # Remove everything and go back to the start of the file
+        opened.seek(0)
+        opened.truncate()
+        # And dump the new file contents
+        json.dump(data, opened, indent=4)
+        # And finally add a new line at the end
+        opened.write("\n")
+
+
 def update_versions():
     """
     Generates the list.json file with the names of the resources.
     """
     # Iterate over the files in the metadata folder
     for file in get_files():
-        # Open it up and load the contents
-        with open(file, "r+") as opened:
-            # Get the information of the resource
-            data = json.load(opened)
-
-            # Create some shortcuts for the objects
-            params = data["update"].get("parameters", {})
-
-            # If the file uses GitHub Releases for the updates
-            if data["update"]["type"] == 1:
-                # Create a new list of versions
-                data["versions"] = []
-
-                # If we need to fetch a specific number of commits
-                if "commits" in params:
-                    # Add those versions from the commits
-                    data["versions"].extend(get_commits(**params))
-
-                # If we don't have the update from releases disabled
-                if not ("skip_releases" in params and params["skip_releases"]):
-                    # Request the list of releases to GitHub and save them
-                    data["versions"] = get_releases(**params)
-
-            # Remove everything and go back to the start of the file
-            opened.seek(0)
-            opened.truncate()
-            # And dump the new file contents
-            json.dump(data, opened, indent=4)
-            # And finally add a new line at the end
-            opened.write("\n")
+        # And update every single one of them
+        update_version(file)
 
 
 def get_files():
